+++ js/technical/temp.js
@@ -1,5 +1,6 @@
 var tmp = {}
 var temp = tmp // Proxy for tmp
+var funcs = {}
 var NaNalert = false;
 
 // Tmp will not call these
@@ -22,8 +23,9 @@ function setupTemp() {
 	tmp.pointGen = {}
 	tmp.displayThings = []
 	tmp.scrolled = 0
+	funcs = {}
 	
-	setupTempData(layers, tmp)
+	setupTempData(layers, tmp, funcs)
 	for (layer in layers){
 		tmp[layer].resetGain = {}
 		tmp[layer].nextAt = {}
@@ -38,7 +40,7 @@ function setupTemp() {
 	temp = tmp
 }
 
-function setupTempData(layerData, tmpData) {
+function setupTempData(layerData, tmpData, funcsData) {
 	for (item in layerData){
 		if (layerData[item] == null) {
 			tmpData[item] = null
@@ -47,16 +49,20 @@ function setupTempData(layerData, tmpData) {
 			tmpData[item] = layerData[item]
 		else if (Array.isArray(layerData[item])) {
 			tmpData[item] = []
-			setupTempData(layerData[item], tmpData[item])
+			funcsData[item] = []
+			setupTempData(layerData[item], tmpData[item], funcsData[item])
 		}
 		else if ((!!layerData[item]) && (layerData[item].constructor === Object)) {
 			tmpData[item] = {}
-			setupTempData(layerData[item], tmpData[item])
+			funcsData[item] = []
+			setupTempData(layerData[item], tmpData[item], funcsData[item])
 		}
 		else if ((!!layerData[item]) && (typeof layerData[item] === "object") && traversableClasses.includes(layerData[item].constructor.name)) {
 			tmpData[item] = new layerData[item].constructor()
+			funcsData[item] = new layerData[item].constructor()
 		}
 		else if (isFunction(layerData[item]) && !activeFunctions.includes(item)){
+			funcsData[item] = layerData[item]
 			tmpData[item] = new Decimal(1) // The safest thing to put probably?
 		} else {
 			tmpData[item] = layerData[item]
@@ -68,7 +74,7 @@ function updateTemp() {
 	if (tmp === undefined)
 		setupTemp()
 
-	updateTempData(layers, tmp)
+	updateTempData(layers, tmp, funcs)
 
 	for (layer in layers){
 		tmp[layer].resetGain = getResetGain(layer)
@@ -95,14 +101,14 @@ function updateTemp() {
 
 }
 
-function updateTempData(layerData, tmpData) {
+function updateTempData(layerData, tmpData, funcsData) {
 	
-	for (item in layerData){
+	for (item in funcsData){
 		if (Array.isArray(layerData[item])) {
-			updateTempData(layerData[item], tmpData[item])
+			updateTempData(layerData[item], tmpData[item], funcsData[item])
 		}
 		else if ((!!layerData[item]) && (layerData[item].constructor === Object) || (typeof layerData[item] === "object") && traversableClasses.includes(layerData[item].constructor.name)){
-			updateTempData(layerData[item], tmpData[item])
+			updateTempData(layerData[item], tmpData[item], funcsData[item])
 		}
 		else if (isFunction(layerData[item]) && !isFunction(tmpData[item])){
 			let value = layerData[item]()
@@ -126,7 +132,7 @@ function updateTempData(layerData, tmpData) {
 
 function updateChallengeTemp(layer)
 {
-	updateTempData(layers[layer].challenges, tmp[layer].challenges)
+	updateTempData(layers[layer].challenges, tmp[layer].challenges, funcs[layer].challenges)
 	updateChallengeDisplay(layer)
 }
 
@@ -144,12 +150,12 @@ function updateChallengeDisplay(layer) {
 
 function updateBuyableTemp(layer)
 {
-	updateTempData(layers[layer].buyables, tmp[layer].buyables)
+	updateTempData(layers[layer].buyables, tmp[layer].buyables, funcs[layer].buyables)
 }
 
 function updateClickableTemp(layer)
 {
-	updateTempData(layers[layer].clickables, tmp[layer].clickables)
+	updateTempData(layers[layer].clickables, tmp[layer].clickables, funcs[layer].clickables)
 }
 
 function constructNodeStyle(layer){